src/bot/requirements.txt
discord.py==1.6.0
python-dotenv==0.15.0
requests==2.25.1